toda aplicao rodando com os 4 arquivos ok
